requirements.txt
PyYAML>=6.0
langchain-core
langchain-openai
python-dotenv
pymysql
matplotlib
